Move CrossSell to public module
No need for it to be hidden only in the -android one
