+++ app/data/data-contract/data-contract-android/src/main/kotlin/com/hedvig/android/data/contract/android/CrossSell.kt
@@ -1,20 +1,6 @@
 package com.hedvig.android.data.contract.android
 
-data class CrossSell(
-  val id: String,
-  val title: String,
-  val subtitle: String,
-  val storeUrl: String,
-  val type: CrossSellType,
-) {
-  enum class CrossSellType {
-    PET,
-    HOME,
-    ACCIDENT,
-    CAR,
-    UNKNOWN,
-  }
-}
+import com.hedvig.android.data.contract.CrossSell
 
 fun CrossSell.CrossSellType.iconRes(): Int = when (this) {
   CrossSell.CrossSellType.PET -> R.drawable.ic_pillow_pet
