+++ app/data/data-contract/data-contract-public/src/main/kotlin/com/hedvig/android/data/contract/CrossSell.kt
@@ -1,0 +1,17 @@
+package com.hedvig.android.data.contract
+
+data class CrossSell(
+  val id: String,
+  val title: String,
+  val subtitle: String,
+  val storeUrl: String,
+  val type: CrossSellType,
+) {
+  enum class CrossSellType {
+    PET,
+    HOME,
+    ACCIDENT,
+    CAR,
+    UNKNOWN,
+  }
+}
