+++ app/ui/cross-sells/src/main/kotlin/com/hedvig/android/crosssells/CrossSells.kt
@@ -36,9 +36,9 @@ import androidx.lifecycle.compose.dropUnlessResumed
 import arrow.core.nonEmptyListOf
 import com.hedvig.android.data.addons.data.TravelAddonBannerInfo
 import com.hedvig.android.data.addons.data.TravelAddonBannerSource
-import com.hedvig.android.data.contract.android.CrossSell
-import com.hedvig.android.data.contract.android.CrossSell.CrossSellType.ACCIDENT
-import com.hedvig.android.data.contract.android.CrossSell.CrossSellType.HOME
+import com.hedvig.android.data.contract.CrossSell
+import com.hedvig.android.data.contract.CrossSell.CrossSellType.ACCIDENT
+import com.hedvig.android.data.contract.CrossSell.CrossSellType.HOME
 import com.hedvig.android.data.contract.android.iconRes
 import com.hedvig.android.design.system.hedvig.ButtonDefaults
 import com.hedvig.android.design.system.hedvig.ButtonDefaults.ButtonSize.Small
@@ -242,12 +242,12 @@ private fun CrossSellItem(
       painter = painterResource(id = type.iconRes()),
       contentDescription = null,
       modifier = Modifier
-        .size(48.dp)
-        .hedvigPlaceholder(
-          visible = isLoading,
-          shape = HedvigTheme.shapes.cornerSmall,
-          highlight = PlaceholderHighlight.fade(),
-        ),
+          .size(48.dp)
+          .hedvigPlaceholder(
+              visible = isLoading,
+              shape = HedvigTheme.shapes.cornerSmall,
+              highlight = PlaceholderHighlight.fade(),
+          ),
     )
     Spacer(Modifier.width(16.dp))
     Column(
